feat: Apply and retry patch files at generation
This change will test patch files up to five times before returning them
for evaluation. If the patch fails to be applied, a minimal patch will
be attempted, followed by a complete regeneration.

This change requires `run_api.py` to provide the additional parameter
`--testbed`, which is the directory that the test repositories are
checked out to. This is where patches will be applied.