fix: Testbed initialization runs on main thread
`TestbedContextManager` doesn't look like it's meant to be run more than
once. Now we'll run it once at the start and use it to distribute
testbed environments to workers. This means each worker operates on a
unique repositories and version pairs. This change also contains a few
bug fixes (introduced by us), necessary to run cleanly.